refactor: move filter and conference reducer out of index
Let us keep the index file clean.
